per review comment: extract else into a function in Autograb.tsx
